pixart-instruct/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
datasets==4.4.1
diffusers==0.35.2
dill==0.4.0
exceptiongroup==1.3.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
packaging==25.0
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
psutil==7.1.3
pyarrow==21.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentencepiece==0.2.1
six==1.17.0
sympy==1.13.1
tokenizers==0.20.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
transformers==4.46.3
triton==3.1.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
